refactor: CoderCoSTEERSettings for finetune
