+++ rdagent/app/finetune/llm/conf.py
@@ -2,16 +2,8 @@ import os
 
 from pydantic_settings import SettingsConfigDict
 
-from rdagent.components.coder.data_science.conf import DSCoderCoSTEERSettings
 from rdagent.core.conf import RD_AGENT_SETTINGS, ExtendedBaseSettings
 from rdagent.scenarios.finetune.utils import prev_model_dirname
-from rdagent.utils.env import (
-    CondaConf,
-    DockerEnv,
-    Env,
-    LLMDockerConf,
-    LocalEnv,
-)
 
 
 class LLMFinetunePropSetting(ExtendedBaseSettings):
@@ -60,37 +52,6 @@ class LLMFinetunePropSetting(ExtendedBaseSettings):
 FT_RD_SETTING = LLMFinetunePropSetting()
 
 
-def get_ft_env(
-    extra_volumes: dict = {},
-    running_timeout_period: int | None = FT_RD_SETTING.debug_timeout,
-    enable_cache: bool | None = None,
-) -> Env:
-    """LLM finetune dedicated environment construction function, equivalent to the responsibility of get_ds_env.
-
-    - Create Docker or Conda environment according to configuration
-    - Configure mount volumes (extra_volumes)
-    - Set running timeout/cache
-    - Call env.prepare() to make the environment ready
-    """
-    conf = DSCoderCoSTEERSettings()
-
-    # Use dedicated LLM docker and conda env
-    if conf.env_type == "docker":
-        env = DockerEnv(conf=LLMDockerConf())
-    elif conf.env_type == "conda":
-        # Use a dedicated llm conda env name
-        env = LocalEnv(conf=CondaConf(conda_env_name="llm_finetune"))
-    else:
-        raise ValueError(f"Unknown env type: {conf.env_type}")
-
-    env.conf.extra_volumes = extra_volumes.copy()
-    env.conf.running_timeout_period = running_timeout_period
-    if enable_cache is not None:
-        env.conf.enable_cache = enable_cache
-    env.prepare()
-    return env
-
-
 def update_settings(
     dataset: str,
     model: str,
@@ -99,7 +60,7 @@ def update_settings(
     Update the RD_AGENT_SETTINGS with the values from LLM_FINETUNE_SETTINGS.
     """
     RD_AGENT_SETTINGS.app_tpl = FT_RD_SETTING.app_tpl
-    os.environ["DS_CODER_COSTEER_EXTRA_EVALUATOR"] = '["rdagent.app.finetune.share.eval.PrevModelLoadEvaluator"]'
+    # Note: extra_evaluator is now configured in FTCoderCoSTEERSettings class
     # Update FT setting instance (use dataset instead of competition in LLM loop)
     FT_RD_SETTING.dataset = dataset
     FT_RD_SETTING.base_model_name = model
