+++ rdagent/scenarios/finetune/train/runner.py
@@ -1,8 +1,8 @@
 from rdagent.app.finetune.llm.conf import FT_RD_SETTING
 from rdagent.components.coder.CoSTEER import CoSTEER
 from rdagent.components.coder.CoSTEER.evaluators import CoSTEERMultiEvaluator
-from rdagent.components.coder.data_science.conf import DSCoderCoSTEERSettings
 from rdagent.components.coder.data_science.share.eval import ModelDumpEvaluator
+from rdagent.components.coder.finetune.conf import FTCoderCoSTEERSettings
 from rdagent.core.scenario import Scenario
 from rdagent.scenarios.data_science.dev.runner import (
     DSRunnerMultiProcessEvolvingStrategy,
@@ -13,7 +13,7 @@ from rdagent.scenarios.finetune.train.eval import (
 )
 
 
-class LLMFinetunePipelineCoSTEERSettings(DSCoderCoSTEERSettings):
+class LLMFinetunePipelineCoSTEERSettings(FTCoderCoSTEERSettings):
     """LLM Fine-tuning specific CoSTEER settings."""
 
     class Config:
